+++ cmoneyw1citymanage/src/Resource.java
@@ -1,5 +1,3 @@
-package cmoneyweek1citymanage;
-
 /**
  * 建造一個起始資源類別
  */
@@ -10,15 +8,12 @@ public class Resource {
     private int woodPeople = 0;//採集木材人數
     private int steelPeople = 0;//採集鋼鐵人數
 
-    private int woodRate = 3;//採集木材效率
-    private int steelRate = 1;//採集鋼鐵效率
-    private int gasRate = 5;//採集瓦斯效率
-
     public Resource(int wood, int steel, int gas) {
         this.gas = gas;
         this.steel = steel;
         this.wood = wood;
     }
+
     public Resource() {
 
     }
@@ -64,32 +59,12 @@ public class Resource {
         this.steelPeople = steelPeople;
     }
 
-    public int getWoodRate() {
-        return woodRate;
-    }
-
-    public void setWoodRate(int woodRate) {
-        this.woodRate = woodRate;
-    }
-
-    public int getSteelRate() {
-        return steelRate;
-    }
-
-    public void setSteelRate(int steelRate) {
-        this.steelRate = steelRate;
-    }
-
-    public int getGasRate() {
-        return gasRate;
-    }
-
-    public void setGasRate(int gasRate) {
-        this.gasRate = gasRate;
-    }
 
     public void addGas(int value) {
         gas += value;
+        if (gas == 0) {
+            System.out.println("已無瓦斯");
+        }
     }
 
     public void addWood(int value) {
@@ -101,6 +76,9 @@ public class Resource {
 
     public void addSteel(int value) {
         steel += value;
+        if (steel == 0) {
+            System.out.println("已無鋼鐵");
+        }
     }
 
     public void addWoodPeople(int value) {
